Refactor profile component: extract form creation and user adaptation logic into separate classes
