+++ front/src/app/components/profile/pages/profile/user.adapter.ts
@@ -1,0 +1,10 @@
+export class UserAdapter {
+  static adapt(apiResponse: any): any {
+    return {
+      name: apiResponse.name ?? '',
+      email: apiResponse.email ?? '',
+      site: apiResponse.site ?? '',
+      phone: apiResponse.phone ?? '',
+    };
+  }
+}
